Log errors, don't return them to the user.
Often times a link will be posted to a site that requires sign-in. LinksBot will report a 403 to the chat which isn't helpful or accurate.
This patch treats those errors as simple warnings and logs them, but does not clutter the chat with it.
This could be made a configuration option, but that doesn't suit my own needs.